requirements.txt
fastapi
uvicorn
monai
pydicom
torchvision
transformers
pandas
openpyxl
nibabel
SimpleITK
matplotlib
numpy
pandas
datasets
import_ipynb